lesson07-cicd-ml-pipeline/pyproject.toml
[project]
name = "lesson07-cicd-ml-pipeline"
version = "0.1.0"
description = "Lesson 07 CI/CD and governance mini-application."
requires-python = ">=3.13,<3.14"
dependencies = [
    "numpy",
    "pandas",
    "pytest",
    "scikit-learn",
    "agent-framework-azure-ai==1.0.0rc3",
    "agent-framework-core==1.0.0rc3",
    "azure-identity",
    "python-dotenv",
]

[tool.uv]
package = false